requirements.txt
argon2-cffi==20.1.0
flask==2.0.1
Flask_SQLAlchemy==2.5.1
floodgate==2.0.0
gunicorn==20.1.0
passlib==1.7.4
psycopg2-binary==2.8.6
pyjwt[crypto]==2.1.0
requests==2.25.1
SQLAlchemy==1.4.17
